stream calling scheme changed, configurations renamed
