+++ machine_translation/__main__.py
@@ -17,8 +17,10 @@ import importlib
 import logging
 import pprint
 
-import config
+import configurations
+
 from machine_translation import main
+from stream import get_tr_stream, get_dev_stream
 
 logger = logging.getLogger(__name__)
 
@@ -30,12 +32,11 @@ parser.add_argument("--bokeh",  default=False, action="store_true",
                     help="Use bokeh server for plotting")
 args = parser.parse_args()
 
-# Get configurations for model
-config = getattr(config, args.proto)()
-
 
 if __name__ == "__main__":
-    logger.info("Model options:\n{}".format(pprint.pformat(config)))
-    data_stream = importlib.import_module(config['stream'])
-    main(config, data_stream.masked_stream, data_stream.dev_stream,
-         args.bokeh)
+    # Get configurations for model
+    configuration = getattr(configurations, args.proto)()
+    logger.info("Model options:\n{}".format(pprint.pformat(configuration)))
+    # Get data streams and call main
+    main(configuration, get_tr_stream(configuration),
+         get_dev_stream(configuration), args.bokeh)
